PR Feedback refactor specs
- Duplicated required feature tests, one with feature flag on and one with feature flag off
- Suffixed with three_choices_inactive to make deleting the tests more straight forward when we remove the feature flag
