+++ spec/system/candidate_interface/apply_again/candidate_applies_again_spec.rb
@@ -1,20 +1,11 @@
 require 'rails_helper'
 
-RSpec.feature 'Apply again' do
+RSpec.feature 'Apply again with three choices' do
   include CandidateHelper
-  include CycleTimetableHelper
 
-  around do |example|
-    Timecop.travel(mid_cycle) do
-      example.run
-    end
-  end
-
-  scenario 'Candidate applies again during mid cycle' do
-    given_the_pilot_is_open
+  scenario 'Candidate applies again with three choices' do
+    given_the_apply_again_with_three_courses_feature_flag_is_active
     and_i_am_signed_in_as_a_candidate
-    and_apply_again_with_three_choices_are_inactive
-
     when_i_have_an_unsuccessful_application
     and_i_visit_the_application_dashboard
     and_i_click_on_apply_again
@@ -26,10 +17,13 @@ RSpec.feature 'Apply again' do
 
     when_i_click_back
     then_i_see_my_current_application_page
-    and_i_am_informed_i_can_only_select_one_course
 
-    when_i_click_through_to_select_a_course
-    then_i_can_indeed_only_select_one_course
+    then_i_should_see_text_suggesting_that_i_can_apply_to_three_courses
+    then_i_can_select_a_course
+    then_i_should_see_text_suggesting_that_i_can_add_two_more_courses
+    and_i_add_two_more_courses
+    then_i_can_see_my_application_with_three_courses
+    and_i_complete_the_section
 
     when_i_complete_my_application
     and_i_skip_feedback
@@ -38,8 +32,8 @@ RSpec.feature 'Apply again' do
     and_i_do_not_see_referee_related_guidance
   end
 
-  def given_the_pilot_is_open
-    FeatureFlag.activate('pilot_open')
+  def given_the_apply_again_with_three_courses_feature_flag_is_active
+    FeatureFlag.activate(:apply_again_with_three_choices)
   end
 
   def and_i_am_signed_in_as_a_candidate
@@ -50,9 +44,7 @@ RSpec.feature 'Apply again' do
   def when_i_have_an_unsuccessful_application
     @application_form = create(
       :completed_application_form,
-      :with_gcses,
       candidate: @candidate,
-      safeguarding_issues_status: :no_safeguarding_issues_to_declare,
       references_completed: true,
     )
     create_list(:selected_reference, 2, application_form: @application_form)
@@ -79,10 +71,6 @@ RSpec.feature 'Apply again' do
     click_link 'First application'
   end
 
-  def then_i_see_the_review_previous_application_page
-    expect(page).to have_current_path(candidate_interface_review_previous_application_path(@application_form.id))
-  end
-
   def when_i_click_back
     click_link 'Back'
   end
@@ -91,15 +79,21 @@ RSpec.feature 'Apply again' do
     expect(page).to have_current_path(candidate_interface_application_form_path)
   end
 
-  def and_i_am_informed_i_can_only_select_one_course
-    expect(page).to have_content('You can only apply to 1 course at a time at this stage of your application')
+  def then_i_see_the_review_previous_application_page
+    expect(page).to have_current_path(candidate_interface_review_previous_application_path(@application_form.id))
   end
 
-  def when_i_click_through_to_select_a_course
-    click_link 'Choose your course', exact: true
+  def then_i_should_see_text_suggesting_that_i_can_apply_to_three_courses
+    expect(page).to have_content('You can apply for up to 3 courses.')
   end
 
-  def then_i_can_indeed_only_select_one_course
+  def and_i_complete_the_section
+    choose 'Yes, I have completed this section'
+    click_button 'Continue'
+  end
+
+  def then_i_can_select_a_course
+    click_link 'Choose your courses', exact: true
     given_courses_exist
 
     choose 'Yes, I know where I want to apply'
@@ -110,15 +104,38 @@ RSpec.feature 'Apply again' do
 
     choose 'Primary (2XT2)'
     click_button t('continue')
+  end
 
-    expect(page).to have_link 'Delete choice'
-    expect(page).to have_content 'I have completed this section'
-    expect(page).not_to have_button 'Add another course'
+  def candidate_submits_application
+    click_link 'Check and submit your application'
+    click_link t('continue')
+
+    # Equality and diversity questions
+    click_link t('continue')
+
+    # What is your sex?
+    choose 'Prefer not to say'
+    click_button t('continue')
+
+    # Are you disabled?
+    choose 'Prefer not to say'
+    click_button t('continue')
+
+    # What is your ethnic group?
+    choose 'Prefer not to say'
+    click_button t('continue')
+
+    # Review page
+    click_link t('continue')
+
+    # Is there anything else you would like to tell us about your application?
+    choose 'No'
+    click_button 'Send application'
+
+    @application = ApplicationForm.last
   end
 
   def when_i_complete_my_application
-    choose t('application_form.completed_radio')
-    click_button t('continue')
     candidate_submits_application
   end
 
@@ -141,7 +158,33 @@ RSpec.feature 'Apply again' do
     expect(page).not_to have_content 'References'
   end
 
-  def and_apply_again_with_three_choices_are_inactive
-    FeatureFlag.deactivate(:apply_again_with_three_choices)
+  def then_i_should_see_text_suggesting_that_i_can_add_two_more_courses
+    expect(page).to have_content('You can choose 2 more courses')
+  end
+
+  def and_i_add_two_more_courses
+    choose 'Yes, add another course'
+    click_button t('continue')
+    choose 'Yes, I know where I want to apply'
+    click_button t('continue')
+    select 'Gorse SCITT (1N1)'
+    click_button t('continue')
+    choose 'Drama (2397)'
+    click_button t('continue')
+
+    choose 'Yes, add another course'
+    click_button t('continue')
+    choose 'Yes, I know where I want to apply'
+    click_button t('continue')
+    select 'Gorse SCITT (1N1)'
+    click_button t('continue')
+    choose 'English (6Z9H)'
+    click_button t('continue')
+  end
+
+  def then_i_can_see_my_application_with_three_courses
+    expect(page).to have_content('Primary (2XT2)')
+    expect(page).to have_content('Drama (2397)')
+    expect(page).to have_content('English (6Z9H)')
   end
 end
